2008-09-10 Weston Schmidt <weston_schmidt@alumni.purdue.edu> 	* src/arguments.c, src/arguments.h, atmel.c, atmel.h 	  commands.c, commands.h, dfu-bool.h, dfu.c, dfu.h 	  main.c: Changed to using dfu_bool & stdint.h values. 	* configure.ac: Removed dependency on stdbool.h.

+++ src/main.c
@@ -38,7 +38,7 @@ int main( int argc, char **argv )
     struct usb_device *device = NULL;
     struct usb_dev_handle *usb_handle = NULL;
     struct programmer_arguments args;
-    unsigned short interface;
+    int32_t interface;
 
     memset( &args, 0, sizeof(args) );
     if( 0 != parse_arguments(&args, argc, argv) ) {
